LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.drawing.style.wordSpacing.measure-expected.txt

2d.text.drawing.style.wordSpacing.measure

Testing word spacing with different length units


FAIL Testing word spacing with different length units assert_approx_equals: word spacing doesn't work. expected 73.9583203125 +/- 0.2 but got 75.5283203125
